# src/main/java/com/edu/web/spacecatsmarket/web/dto/category/ResponseCategoryDto.java
package com.edu.web.spacecatsmarket.web.dto.category;

import lombok.Builder;
import lombok.extern.jackson.Jacksonized;

@Builder
@Jacksonized
public record ResponseCategoryDto(
        String id,
        String name
) {}